Moved base resources to be more consistent. Moved services.py while I was there.
