--- ScoutSuite/providers/base/configs/__init__.py
@@ -1,7 +1,0 @@
-
-resource_id_map = {
-    'network_interfaces': 'NetworkInterfaceId',
-    'peering_connections': 'VpcPeeringConnectionId',
-    'subnet_groups': 'DBSubnetGroupName'
-}
-
