+++ ScoutSuite/providers/base/resources/base.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 This module provides some abstract classes for representing a hierarchical structure.
 Especially since all cloud providers (AWS, Azure and GCP for now) organize their resources (virtual machines,
@@ -34,6 +33,9 @@ class CompositeResources(Resources, metaclass=abc.ABCMeta):
     referred to as its 'children'.
     """
 
+    def __init__(self, facade):
+        self.facade = facade
+
     @property
     @abc.abstractmethod
     def _children(self):
@@ -57,7 +59,8 @@ class CompositeResources(Resources, metaclass=abc.ABCMeta):
 
         tasks = {
             asyncio.ensure_future(
-                self._fetch_children(resource_parent=resource_parent, scope=scopes[resource_parent_key])
+                self._fetch_children(
+                    resource_parent=resource_parent, scope=scopes[resource_parent_key])
             ) for (resource_parent_key, resource_parent) in resources.items()
         }
         await asyncio.wait(tasks)
@@ -70,10 +73,12 @@ class CompositeResources(Resources, metaclass=abc.ABCMeta):
         :param resource_parent: The resource in which the children will be stored.
         :param scope: The scope passed to the children constructors.
         """
-        children = [(child_class(self.facade, **scope), child_name) for (child_class, child_name) in self._children]
+        children = [(child_class(self.facade, **scope), child_name)
+                    for (child_class, child_name) in self._children]
         # Fetch all children concurrently:
         await asyncio.wait(
-            {asyncio.ensure_future(child.fetch_all()) for (child, _) in children}
+            {asyncio.ensure_future(child.fetch_all())
+             for (child, _) in children}
         )
         # Update parent content:
         for child, child_name in children:
