+++ ScoutSuite/providers/base/services.py
