+++ ScoutSuite/providers/gcp/resources/base.py
@@ -1,7 +1,7 @@
 """This module provides implementations for CompositeResources for GCP."""
 
 import asyncio
-from ScoutSuite.providers.base.configs.resources import CompositeResources
+from ScoutSuite.providers.base.resources.base import CompositeResources
 
 class GCPCompositeResources(CompositeResources):
