+++ ScoutSuite/providers/gcp/resources/cloudresourcemanager/bindings.py
@@ -1,5 +1,5 @@
 from ScoutSuite.core.console import print_exception
-from ScoutSuite.providers.base.configs.resources import Resources
+from ScoutSuite.providers.base.resources.base import Resources
 from ScoutSuite.providers.gcp.facade.gcp import GCPFacade
 from ScoutSuite.providers.utils import get_non_provider_id
