+++ ScoutSuite/providers/gcp/resources/gce/disks.py
@@ -1,4 +1,4 @@
-from ScoutSuite.providers.base.configs.resources import Resources
+from ScoutSuite.providers.base.resources.base import Resources
 from ScoutSuite.providers.utils import get_non_provider_id
 
 class Disks(Resources):
