+++ ScoutSuite/providers/gcp/resources/iam/bindings.py
@@ -1,5 +1,5 @@
 from ScoutSuite.providers.gcp.facade.gcp import GCPFacade
-from ScoutSuite.providers.base.configs.resources import Resources
+from ScoutSuite.providers.base.resources.base import Resources
 
 class Bindings(Resources):
     def __init__(self, facade: GCPFacade, project_id: str, service_account_email: str):
